pyproject.toml
[project]
name = "crewaiProject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai-tools>=0.55.0",
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-ollama>=0.3.5",
    "linkup>=0.1.4",
    "linkup-sdk>=0.2.8",
    "streamlit>=1.47.0",
    "uvicorn>=0.35.0",
]
